chore(TCOMP-681): move TacokitElementParameter to appropriate package
