+++ component-studio-integration/src/main/java/org/talend/sdk/component/studio/ui/composite/TaCoKitWizardComposite.java
@@ -17,10 +17,10 @@ import org.eclipse.swt.widgets.Composite;
 import org.talend.commons.exception.ExceptionHandler;
 import org.talend.core.model.process.EComponentCategory;
 import org.talend.core.model.process.Element;
-import org.talend.sdk.component.studio.metadata.TaCoKitElementParameter;
-import org.talend.sdk.component.studio.metadata.TaCoKitElementParameter.IValueChangedListener;
 import org.talend.sdk.component.studio.metadata.model.TaCoKitConfigurationModel;
 import org.talend.sdk.component.studio.metadata.model.TaCoKitConfigurationModel.ValueModel;
+import org.talend.sdk.component.studio.model.parameter.TaCoKitElementParameter;
+import org.talend.sdk.component.studio.model.parameter.TaCoKitElementParameter.IValueChangedListener;
 
 /**
  * Stores {@link TaCoKitConfigurationModel} and sets Configuration Model Updater listener for each
